+++ src/flask_nova/route_refactor.py
@@ -3,47 +3,45 @@ import inspect as ip
 import functools as ft
 import typing as t
 from flask import request
-
 from .utils import filter_options
 from .responses import _bind_route_parameters, ResponseSerializer
 from . import types as nt
-
-
 from .di import resolve_dependencies
+
 
 P = t.ParamSpec("P")
 R = t.TypeVar("R")
 
+
 class RouteFactory:
-    def __init__(self, serializer: ResponseSerializer)-> None:
+    def __init__(self, serializer: ResponseSerializer) -> None:
         self.serializer = serializer
 
     def build(
-            self,
-            owner,
-            rule: str,
-            methods: t.List[nt.Method],
-            tags: t.Optional[t.List[t.Union[str, Enum]]],
-            response_model: t.Any| None,
-            summary: t.Optional[str]= None,
-            description: t.Optional[str]= None,
-            provide_automatic_options: bool | None = None,
-            **options: t.Dict[str, t.Any])-> t.Callable[[nt.T_route], nt.T_route]:
+        self,
+        owner,
+        rule: str,
+        methods: t.List[nt.Method],
+        tags: t.Optional[t.List[t.Union[str, Enum]]],
+        response_model: t.Any | None,
+        summary: t.Optional[str] = None,
+        description: t.Optional[str] = None,
+        provide_automatic_options: bool | None = None,
+        **options: t.Dict[str, t.Any],
+    ) -> t.Callable[[nt.T_route], nt.T_route]:
 
-
-        def decorator(func: nt.T_route)->  nt.T_route:
-            is_async = ip.iscoroutinefunction(func)
+        def decorator(func: nt.T_route) -> nt.T_route:
             sig = ip.signature(func)
             type_hints = t.get_type_hints(func)
 
             f = resolve_dependencies(func)
-
 
             setattr(f, "_flasknova_tags", tags or [])
             setattr(f, "_flasknova_response_model", response_model)
             setattr(f, "_flasknova_summary", summary)
             setattr(f, "_flasknova_description", description)
 
+            # todo: add servers,  mermaid js and responses to the route param
 
             @ft.wraps(f)
             async def handler(*args, **kwargs):
@@ -52,7 +50,7 @@ class RouteFactory:
                     return bound_values
 
                 try:
-                    result = await f(**bound_values) 
+                    result = await f(**bound_values)
                 except Exception as e:
                     raise e