+++ tests/conftest.py
@@ -1,0 +1,19 @@
+import pytest
+from flask_nova import FlaskNova
+
+
+@pytest.fixture
+def app():
+    app = FlaskNova(__name__)
+    app.config.update({"TESTING": True})
+    yield app
+
+
+@pytest.fixture()
+def client(app):
+    return app.test_client()
+
+
+@pytest.fixture
+def runner(app):
+    app.test_cli_runner()